Use the DynamicEventDispatcher to build a super lightweight projector.
Added documentation.


Use a ReadOnlySet to be able to quickly determine if the dispatcher supports an event type.

Descriptive method name.

drop project.

Update coverage.

clean-up.
